refactor: simplify blocks parsing code
