+++ site/src/lib/blocks.ts
@@ -1,4 +1,5 @@
-import { BlockMetadata } from "blockprotocol";
+import { BlockMetadata, BlockMetadataRepository } from "blockprotocol";
+import hostedGitInfo from "hosted-git-info";
 
 /** @todo type as JSON object */
 export type BlockProps = object;
@@ -6,6 +7,7 @@ export type BlockProps = object;
 export type ExpandedBlockMetadata = BlockMetadata & {
   packagePath: string;
   lastUpdated?: string;
+  repository?: string;
 };
 
 export type BuildConfig = {
@@ -30,6 +32,34 @@ const getBlockMediaUrl = (
 
   return `/blocks/${packagePath}/${mediaPath}`;
 };
+
+// this only runs on the server-side because hosted-git-info uses some nodejs dependencies
+const getRepositoryUrl = (
+  repository: BlockMetadataRepository | null,
+): string | null => {
+  if (typeof repository === "string") {
+    const repositoryUrl = hostedGitInfo.fromUrl(repository)?.browse("");
+
+    if (repositoryUrl) {
+      return repositoryUrl;
+    }
+
+    return null;
+  }
+
+  const { url, directory } = repository ?? {};
+
+  if (url) {
+    const repositoryUrl = hostedGitInfo.fromUrl(url)?.browse(directory ?? "");
+
+    if (repositoryUrl) {
+      return repositoryUrl;
+    }
+  }
+
+  return null;
+};
+
 /**
  * used to read block metadata from disk.
  *
@@ -40,29 +70,30 @@ export const readBlocksFromDisk = (): ExpandedBlockMetadata[] => {
   const glob = require("glob");
   /* eslint-enable global-require */
 
-  const buildConfig: BuildConfig[] = glob
-    .sync(`${process.cwd()}/../hub/**/*.json`)
-    .map((path: string) => ({
-      ...JSON.parse(fs.readFileSync(path, { encoding: "utf8" })),
-    }));
-
   return glob
     .sync(`${process.cwd()}/public/blocks/**/block-metadata.json`)
-    .map((path: string) => ({
-      // @todo should be redundant to block's package.json#name
-      packagePath: path.split("/").slice(-3, -1).join("/"),
-      ...JSON.parse(fs.readFileSync(path, { encoding: "utf8" })),
-    }))
-    .map((metadata: ExpandedBlockMetadata) => {
-      const blockConfig = buildConfig.find(
-        ({ workspace }) => workspace === metadata.name,
+    .map((path: string) => {
+      const packagePath = path.split("/").slice(-3, -1).join("/");
+
+      const metadata: ExpandedBlockMetadata = {
+        // @todo should be redundant to block's package.json#name
+        packagePath,
+        ...JSON.parse(fs.readFileSync(path, { encoding: "utf8" })),
+      };
+
+      const blockConfig: BuildConfig = JSON.parse(
+        fs.readFileSync(`${process.cwd()}/../hub/${packagePath}.json`, {
+          encoding: "utf8",
+        }),
       );
 
-      let repository = metadata.repository ?? null;
+      let blockRepository = metadata.repository ?? null;
 
-      if (!repository && blockConfig) {
-        repository = `${blockConfig.repository}#${blockConfig.branch}`;
+      if (!blockRepository && blockConfig) {
+        blockRepository = `${blockConfig.repository}#${blockConfig.branch}`;
       }
+
+      const repository = getRepositoryUrl(blockRepository);
 
       return {
         ...metadata,
